fix: Extracted inner interface
