+++ src/main/java/com/redis/smartcache/cli/components/AbstractTableSelectorComponent.java
@@ -1,13 +1,7 @@
 package com.redis.smartcache.cli.components;
 
-import org.jline.keymap.BindingReader;
-import org.jline.keymap.KeyMap;
-import org.jline.terminal.Terminal;
-import org.jline.utils.InfoCmp;
-import org.springframework.shell.component.context.BaseComponentContext;
-import org.springframework.shell.component.context.ComponentContext;
-import org.springframework.shell.component.support.*;
-import org.springframework.util.ObjectUtils;
+import static org.jline.keymap.KeyMap.ctrl;
+import static org.jline.keymap.KeyMap.key;
 
 import java.util.Comparator;
 import java.util.List;
@@ -16,10 +10,20 @@ import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-import static org.jline.keymap.KeyMap.ctrl;
-import static org.jline.keymap.KeyMap.key;
+import org.jline.keymap.BindingReader;
+import org.jline.keymap.KeyMap;
+import org.jline.terminal.Terminal;
+import org.jline.utils.InfoCmp;
+import org.springframework.shell.component.context.BaseComponentContext;
+import org.springframework.shell.component.support.AbstractComponent;
+import org.springframework.shell.component.support.Enableable;
+import org.springframework.shell.component.support.Itemable;
+import org.springframework.shell.component.support.Matchable;
+import org.springframework.shell.component.support.Nameable;
+import org.springframework.shell.component.support.Selectable;
+import org.springframework.util.ObjectUtils;
 
-public abstract class AbstractTableSelectorComponent<T, C extends AbstractTableSelectorComponent.SelectorComponentContext<T, I, C>, I extends Nameable & Matchable & Enableable & Selectable & Itemable<T>>
+public abstract class AbstractTableSelectorComponent<T, C extends SelectorComponentContext<T, I, C>, I extends Nameable & Matchable & Enableable & Selectable & Itemable<T>>
         extends AbstractComponent<C> {
 
     protected final String name;
@@ -252,90 +256,6 @@ public abstract class AbstractTableSelectorComponent<T, C extends AbstractTableS
         }
     }
 
-    /**
-     * Context interface on a selector component sharing content.
-     */
-    public interface SelectorComponentContext<T, I extends Nameable & Matchable & Itemable<T>, C extends SelectorComponentContext<T, I, C>>
-            extends ComponentContext<C> {
-
-        /**
-         * Sets a name
-         *
-         * @param name the name
-         */
-        void setName(String name);
-
-        /**
-         * Gets an input.
-         *
-         * @return an input
-         */
-        String getInput();
-
-        /**
-         * Gets an item states
-         *
-         * @return an item states
-         */
-        List<ItemState<I>> getItemStates();
-
-        /**
-         * Sets an item states.
-         *
-         * @param itemStateView the input state
-         */
-        void setItemStates(List<ItemState<I>> itemStateView);
-
-        /**
-         * Sets an item state view
-         *
-         * @param itemStateView the item state view
-         */
-        void setItemStateView(List<ItemState<I>> itemStateView);
-
-        /**
-         * Gets a cursor row.
-         *
-         * @return a cursor row.
-         */
-        Integer getCursorRow();
-
-        /**
-         * Sets a cursor row.
-         *
-         * @param cursorRow the cursor row
-         */
-        void setCursorRow(Integer cursorRow);
-
-        /**
-         * Gets an items.
-         *
-         * @return an items
-         */
-        List<I> getItems();
-
-        /**
-         * Sets an items.
-         *
-         * @param items the items
-         */
-        void setItems(List<I> items);
-
-        /**
-         * Gets a result items.
-         *
-         * @return a result items
-         */
-        List<I> getResultItems();
-
-        /**
-         * Sets a result items.
-         *
-         * @param items the result items
-         */
-        void setResultItems(List<I> items);
-
-    }
 
     /**
      * Base implementation of a {@link org.springframework.shell.component.support.AbstractSelectorComponent.SelectorComponentContext}.
@@ -343,6 +263,7 @@ public abstract class AbstractTableSelectorComponent<T, C extends AbstractTableS
     protected static class BaseSelectorComponentContext<T, I extends Nameable & Matchable & Itemable<T>, C extends SelectorComponentContext<T, I, C>>
             extends BaseComponentContext<C> implements SelectorComponentContext<T, I, C> {
 
+        private static final long serialVersionUID = 1L;
         private String name;
         private String input;
         private List<ItemState<I>> itemStates;
